testing/web-platform/tests/workers/WorkerLocation_pathname.htm

<!DOCTYPE html>
<html>
<head>
<title> WorkerLocation URL decomposition IDL attribute: pathname </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id=log></div>
<script>
    
    var FileName = './support/WorkerLocation.js';
    var pathname = window.location.pathname;
    var ExpectedResult = pathname.substring(0, pathname.lastIndexOf('/')) + FileName.substring(1);
    
    var t = async_test("Test Description: WorkerLocation pathname attribute returns the current path in the underlying URL.");
                  
    var worker = new Worker(FileName);

    worker.onmessage = t.step_func(function(evt)
    {
        assert_equals(evt.data.pathname, ExpectedResult);
        t.done();
    });
</script>
</body>
</html>
